requirements.txt
# Helm dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.9

# Desktop control
pyautogui>=0.9.54
pyperclip>=1.8.2
mss>=9.0.1
Pillow>=10.0.0
pywin32>=306

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# LLM clients
anthropic>=0.25.0
google-generativeai>=0.5.0
openai>=1.14.0

# Scheduling
apscheduler>=3.10.4

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
aiofiles>=23.2.1
httpx>=0.27.0
